fix(git): Correct git_branch_state function display
When called while the current git HEAD is attached to one single tag,
this function was outputting "<TAG_NAME>," (with an extra comma).
This commit fixes this behaviour and handles an old TODO related to a
git command alias set when no external prompt solution (like bash-it,
oh-my-bash, oh-my-zsh, ...) is enabled.